Started ToneCreator

+++ include/NoteTone.h
@@ -16,13 +16,13 @@ class NoteTone: public Tone{
 		/* Must be an int from 1 to 32. EndBeatPosition is the beat position where the tone will be played in a 2 measure interval.
 		 * Measures are always in 4/4 time with each measure containing 16 beats, making 2 measures contain 32 beats. */
 		WaveForm waveform;
-		int frequency;
+		float frequency;
 
 	public:
-		NoteTone(unsigned short int endBeat, unsigned short int startBeat, WaveForm waveform, int frequency);
+		NoteTone(unsigned short int endBeat, unsigned short int startBeat, WaveForm waveform, float frequency);
 		unsigned short int getEndBeatPosition();
 		WaveForm getWaveform();
-		int getFrequency();
+		float getFrequency();
 };
 
 #endif
